refactor: create useFileUploadZoneLogic
